Revert "refactor: Moved dataclasses to their own module"
This reverts commit 624138de670c86b8923e82803c432c34a618c7e0.
